requirements.txt
requests
beautifulsoup4
lxml
supabase
python-dotenv
psycopg2-binary
tqdm
